refactor: split buzz/api into domain packages
buzz/api/__init__.py held 1364 lines covering booking, tickets,
sponsorships, check-in, account, campaigns and coupons in one module.
Since a Frappe endpoint's URL is its dotted module path, that layout also
made the public API surface unreadable.

Each domain now owns a package, so an endpoint's URL names its domain:
buzz.api.booking.process_booking rather than buzz.api.process_booking.
forms.py, auth.py and proposals.py become packages of the same shape.

Function bodies are unchanged; this is relocation only. Landed alongside
are api/schemas.py and api/exceptions.py, the pydantic and error base
classes later PRs build on, currently unused.
