+++ buzz/ticketing/doctype/buzz_coupon_code/test_buzz_coupon_code.py
@@ -1099,7 +1099,7 @@ class TestValidateCouponAPI(IntegrationTestCase):
 
 	def test_validate_coupon_returns_discount_info(self):
 		"""Test that validate_coupon returns correct discount info."""
-		from buzz.api import validate_coupon
+		from buzz.api.booking import validate_coupon
 
 		frappe.get_doc(
 			{
@@ -1121,7 +1121,7 @@ class TestValidateCouponAPI(IntegrationTestCase):
 
 	def test_validate_coupon_returns_max_and_min_values(self):
 		"""Test that validate_coupon returns max_discount_amount and min_order_value."""
-		from buzz.api import validate_coupon
+		from buzz.api.booking import validate_coupon
 
 		frappe.get_doc(
 			{
@@ -1144,7 +1144,7 @@ class TestValidateCouponAPI(IntegrationTestCase):
 
 	def test_validate_coupon_returns_free_tickets_info(self):
 		"""Test that validate_coupon returns correct free tickets info."""
-		from buzz.api import validate_coupon
+		from buzz.api.booking import validate_coupon
 
 		frappe.get_doc(
 			{
@@ -1167,7 +1167,7 @@ class TestValidateCouponAPI(IntegrationTestCase):
 
 	def test_validate_coupon_invalid_code(self):
 		"""Test that invalid coupon code returns error."""
-		from buzz.api import validate_coupon
+		from buzz.api.booking import validate_coupon
 
 		result = validate_coupon("INVALIDCODE", str(self.test_event.name))
 
@@ -1178,7 +1178,7 @@ class TestValidateCouponAPI(IntegrationTestCase):
 
 	def test_coupon_not_yet_active(self):
 		"""Test that coupon with future valid_from is rejected."""
-		from buzz.api import validate_coupon
+		from buzz.api.booking import validate_coupon
 
 		frappe.get_doc(
 			{
@@ -1199,7 +1199,7 @@ class TestValidateCouponAPI(IntegrationTestCase):
 
 	def test_expired_coupon_rejected(self):
 		"""Test that expired coupon is rejected."""
-		from buzz.api import validate_coupon
+		from buzz.api.booking import validate_coupon
 
 		frappe.get_doc(
 			{
@@ -1220,7 +1220,7 @@ class TestValidateCouponAPI(IntegrationTestCase):
 
 	def test_coupon_within_validity_period(self):
 		"""Test that coupon within valid date range works."""
-		from buzz.api import validate_coupon
+		from buzz.api.booking import validate_coupon
 
 		frappe.get_doc(
 			{
@@ -1243,7 +1243,7 @@ class TestValidateCouponAPI(IntegrationTestCase):
 
 	def test_max_usage_per_user_enforced(self):
 		"""Test that user cannot exceed per-user usage limit."""
-		from buzz.api import validate_coupon
+		from buzz.api.booking import validate_coupon
 
 		coupon = frappe.get_doc(
 			{
@@ -1283,7 +1283,7 @@ class TestValidateCouponAPI(IntegrationTestCase):
 
 	def test_per_user_limit_does_not_affect_other_users(self):
 		"""Test that per-user limit doesn't block other users."""
-		from buzz.api import validate_coupon
+		from buzz.api.booking import validate_coupon
 
 		coupon = frappe.get_doc(
 			{
