+++ buzz/ticketing/doctype/event_booking/test_event_booking.py
@@ -471,7 +471,7 @@ class TestProcessBookingAPI(IntegrationTestCase):
 
 	def test_process_booking_with_utm_parameters(self):
 		"""Test that process_booking API correctly saves UTM parameters."""
-		from buzz.api import process_booking
+		from buzz.api.booking import process_booking
 
 		test_event = frappe.get_doc("Buzz Event", {"route": "test-route"})
 
@@ -826,7 +826,7 @@ class TestProcessBookingAPI(IntegrationTestCase):
 
 	def test_process_booking_offline_stays_in_draft(self):
 		"""Test that offline bookings via process_booking stay in draft with no tickets."""
-		from buzz.api import process_booking
+		from buzz.api.booking import process_booking
 
 		test_event = frappe.get_doc("Buzz Event", {"route": "test-route"})
 		test_event.apply_tax = False
@@ -883,7 +883,7 @@ class TestProcessBookingAPI(IntegrationTestCase):
 
 	def test_process_booking_offline_generates_tickets_on_approval(self):
 		"""Test that approving an offline booking created via API generates tickets."""
-		from buzz.api import process_booking
+		from buzz.api.booking import process_booking
 
 		test_event = frappe.get_doc("Buzz Event", {"route": "test-route"})
 		test_event.apply_tax = False
@@ -945,7 +945,7 @@ class TestProcessBookingAPI(IntegrationTestCase):
 	def test_process_booking_free_event_returns_redirect_to(self):
 		"""Free bookings (total_amount == 0) must get a redirect_to the new
 		token-gated booking-success screen, same as the paid-gateway branch."""
-		from buzz.api import process_booking, verify_booking_access_token
+		from buzz.api.booking import process_booking, verify_booking_access_token
 
 		test_event = frappe.get_doc("Buzz Event", {"route": "test-route"})
 		test_event.apply_tax = False
@@ -983,7 +983,7 @@ class TestProcessBookingAPI(IntegrationTestCase):
 
 	def test_process_booking_without_utm_parameters(self):
 		"""Test that process_booking API works without UTM parameters."""
-		from buzz.api import process_booking
+		from buzz.api.booking import process_booking
 
 		test_event = frappe.get_doc("Buzz Event", {"route": "test-route"})
 		test_event.is_published = True
@@ -1025,7 +1025,7 @@ class TestProcessBookingAPI(IntegrationTestCase):
 
 	def test_process_booking_with_empty_utm_parameters(self):
 		"""Test that process_booking API handles empty UTM list."""
-		from buzz.api import process_booking
+		from buzz.api.booking import process_booking
 
 		test_event = frappe.get_doc("Buzz Event", {"route": "test-route"})
 
@@ -1066,7 +1066,7 @@ class TestProcessBookingAPI(IntegrationTestCase):
 
 	def test_process_booking_failed_for_unpublished_event(self):
 		"""Booking must fail when Buzz Event is not published."""
-		from buzz.api import process_booking
+		from buzz.api.booking import process_booking
 
 		test_event = frappe.get_doc("Buzz Event", {"route": "test-route"})
 		test_event.is_published = 0
@@ -1165,7 +1165,7 @@ class TestBookingConfirmation(IntegrationTestCase):
 		return booking
 
 	def test_access_token_roundtrip(self):
-		from buzz.api import get_booking_access_token, verify_booking_access_token
+		from buzz.api.booking import get_booking_access_token, verify_booking_access_token
 
 		token = get_booking_access_token("B-TEST-001")
 		self.assertTrue(verify_booking_access_token("B-TEST-001", token))
@@ -1178,7 +1178,7 @@ class TestBookingConfirmation(IntegrationTestCase):
 		self.assertFalse(verify_booking_access_token("B-OTHER-002", token))
 
 	def test_get_booking_confirmation_valid_token_as_guest(self):
-		from buzz.api import get_booking_access_token, get_booking_confirmation
+		from buzz.api.booking import get_booking_access_token, get_booking_confirmation
 
 		booking = self._make_submitted_booking()
 		token = get_booking_access_token(booking.name)
@@ -1195,7 +1195,7 @@ class TestBookingConfirmation(IntegrationTestCase):
 		self.assertTrue(result["event"]["title"])
 
 	def test_get_booking_confirmation_bad_token_as_guest_raises(self):
-		from buzz.api import get_booking_confirmation
+		from buzz.api.booking import get_booking_confirmation
 
 		booking = self._make_submitted_booking()
 
@@ -1209,7 +1209,7 @@ class TestBookingConfirmation(IntegrationTestCase):
 			frappe.set_user("Administrator")
 
 	def test_get_booking_confirmation_owner_without_token(self):
-		from buzz.api import get_booking_confirmation
+		from buzz.api.booking import get_booking_confirmation
 
 		booking = self._make_submitted_booking()
 
@@ -1400,7 +1400,7 @@ class TestZoomBackedCategoryBooking(IntegrationTestCase):
 		frappe.db.rollback()
 
 	def _book_without_last_name(self, category):
-		from buzz.api import process_booking
+		from buzz.api.booking import process_booking
 
 		self.event.db_set("category", category)
 		return process_booking(
